+++ src/main/kotlin/com/github/igr/sexyeditor/ui/PluginIcons.kt
@@ -4,10 +4,6 @@ import javax.swing.ImageIcon
 
 class PluginIcons {
     companion object {
-        val MOVE_DOWN = actionIconOf("moveDown")
-        val MOVE_UP = actionIconOf("moveUp")
-        val REMOVE = generalIconOf("remove")
-        val ADD = generalIconOf("add")
         val IMAGES = iconOf("images")
         val POSITIONS = arrayOf(
             iconOf("top-left"),
@@ -27,18 +23,6 @@ class PluginIcons {
 
 private fun iconOf(name: String): ImageIcon {
     val path = "/icons/$name.png"
-    val url = PluginIcons::class.java.getResource(path)
-    return ImageIcon(url)
-}
-
-private fun generalIconOf(name: String): ImageIcon {
-    val path = "/general/$name.png"
-    val url = PluginIcons::class.java.getResource(path)
-    return ImageIcon(url)
-}
-
-private fun actionIconOf(name: String): ImageIcon {
-    val path = "/actions/$name.png"
     val url = PluginIcons::class.java.getResource(path)
     return ImageIcon(url)
 }